fix double->int

+++ lib/data/sources/network/api/v2_api.dart
@@ -1506,7 +1506,7 @@ class SendTxParams {
   final String source;
   final String destination;
   final String asset;
-  final double quantity;
+  final int quantity;
   final String? memo;
   final bool memoIsHex;
   final bool useEnhancedSend;
@@ -2297,7 +2297,7 @@ abstract class V2Api {
     @Path("address") String address,
     @Query("destination") String destination,
     @Query("asset") String asset,
-    @Query("quantity") double quantity, [
+    @Query("quantity") int quantity, [
     @Query("allow_unconfirmed_inputs") bool? allowUnconfirmedInputs,
     @Query("fee") int? fee,
   ]);
@@ -2308,7 +2308,7 @@ abstract class V2Api {
     @Path("address") String address,
     @Query("destination") String destination,
     @Query("asset") String asset,
-    @Query("quantity") double quantity, [
+    @Query("quantity") int quantity, [
     @Query("allow_unconfirmed_inputs") bool? allowUnconfirmedInputs,
     @Query("fee") int? fee,
   ]);
